@@ -56,7 +56,7 @@
     "@babel/preset-env": "7.14.2",
     "@babel/preset-react": "7.13.13",
     "@babel/preset-typescript": "7.13.0",
-    "autoprefixer": "10.2.5",
+    "autoprefixer": "10.2.6",
     "babel-loader": "8.2.2",
     "babel-plugin-transform-class-properties": "6.24.1",
     "cross-env": "7.0.3",
@@ -66,9 +66,9 @@
     "postcss-loader": "5.3.0",
     "sass-loader": "11.1.1",
     "terser-webpack-plugin": "5.1.2",
-    "ts-loader": "8.1.0",
-    "typescript": "4.2.4",
-    "webpack": "5.37.1",
+    "ts-loader": "9.2.3",
+    "typescript": "4.3.2",
+    "webpack": "5.38.1",
     "webpack-bundle-analyzer": "4.4.1",
     "webpack-cli": "4.7.0"
   },
