@@ -36,10 +36,10 @@
   "devDependencies": {
     "@babel/core": "7.14.3",
     "@babel/plugin-transform-arrow-functions": "7.13.0",
-    "@babel/plugin-transform-async-to-generator": "7.13.0",
+    "@babel/plugin-transform-async-to-generator": "7.14.5",
     "@babel/plugin-transform-runtime": "^7.12.10",
-    "@babel/plugin-transform-typescript": "7.14.3",
-    "@babel/preset-env": "7.14.1",
+    "@babel/plugin-transform-typescript": "7.14.5",
+    "@babel/preset-env": "7.14.5",
     "@babel/preset-typescript": "7.13.0",
     "babel-loader": "8.2.2",
     "babel-plugin-transform-class-properties": "6.24.1",
@@ -47,8 +47,8 @@
     "css-loader": "5.2.6",
     "file-loader": "6.2.0",
     "sass-loader": "11.1.1",
-    "ts-loader": "8.1.0",
-    "webpack": "5.37.1",
+    "ts-loader": "9.2.3",
+    "webpack": "5.38.1",
     "webpack-bundle-analyzer": "4.4.2",
     "webpack-cli": "4.7.0"
   },
